Index.html and chartify sample text
Wrote some hodgepodge sample text for index.html, although I want to
create a sort of step by step question guide in the future (“which of
the following do you already have?”: “a written pattern”; “an idea”…
etc). This is just a placeholder.

Also put in an (actual) placeholder for the textarea in chartify.