Implemented world storage and window resize
Not too efficient, may come back to storage optimization if it seems to
be a problem once the stitches are done.

Copied in window resize event listener from chartify. May not want to
do all of the default and parse upon resize, but that can be decided
with optimization as well.