event listeners for user wpi and hook sizes
just simple html selects for now, but correctly changes sizes within
all stitches.

may want to scale the size of the tail to the yarn radius in some way,
but that’s not too important. May come back to it if it looks too
strange for the smaller hook/yarn patterns